Move person processing / person profile logic out of the stateless interface
Looking at other server-side SDKs, these don't look to be considered.
